+++ src/library/Pqi/Conformance/Operation/Connectdb.hs
@@ -14,6 +14,7 @@ import qualified Data.Text as Text
 import qualified Pqi
 import Pqi.Conformance.Harness
 import Pqi.Conformance.Observation
+import qualified Pqi.Conformance.Operation.Connectdb.HandshakeReset as HandshakeReset
 import qualified Pqi.Conformance.Operation.Connectdb.MissingUnixSocketDirectory as MissingUnixSocketDirectory
 import qualified Pqi.Conformance.Operation.Connectdb.Rejection as Rejection
 import qualified Pqi.Conformance.Operation.Connectdb.UnixSocketUri as UnixSocketUri
@@ -29,6 +30,7 @@ spec adapter = do
   UnixSocketUri.spec adapter
   MissingUnixSocketDirectory.spec adapter
   UnresolvableHost.spec adapter
+  HandshakeReset.spec adapter
 
   describe "connectdb" do
     it "opens a usable connection" \conninfo ->